+++ src/test/java/kitchenpos/fakeobject/FakeProfanityClient.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package kitchenpos.fakeobject;
 
 import kitchenpos.infra.ProfanityClient;
